Add CFA and SEM models for Feelings of Preparedness
I run a CFA to confirm the factor structure for Feelings of Preparedness
and an SEM model to calculate the effects of using GitHub in the
classroom on students' feeling of preparedness.

I find that students in classroom where GitHub was used feel 0.30
standards deviations more prepared that students in classrooms where
GitHub was not used.